% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuitsdata.R
\docType{data}
\name{circuits_data}
\alias{circuits_data}
\title{Circuits Dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 77 rows and 9 columns.
}
\usage{
circuits_data
}
\description{
This dataset includes the links to render the circuit images for the app
}
\keyword{datasets}
