% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driversmemes.R
\docType{data}
\name{drivers_memes}
\alias{drivers_memes}
\title{Drivers Dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 29 rows and 11 columns.
}
\usage{
drivers_memes
}
\description{
This dataset includes the links to render the driver images for the app
}
\keyword{datasets}
