% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selectYear}
\alias{selectYear}
\title{Select Year}
\usage{
selectYear(inputId, label, years = c("2019", "2020", "2021", "2022", "2023"))
}
\arguments{
\item{inputId}{input slot that will be used to access the value}

\item{label}{label for the selected input}

\item{years}{vector of years to choose from: c("2019", "2020", "2021", "2022", "2023")}
}
\value{
enables user to choose a year from a drop-down menu
}
\description{
this function selects years 2019, 2020, 2021, 2022 and 2023 out of the years between 1950-2023
}
